Feat/v2.1.0 (#119)
* fix: ai comment && accept disc can not delete

* fix: ai comment

* refactor: clean up Forum component and enhance user trend descriptions

- Removed commented-out code related to issue category selection in the Forum component for better readability.
- Added a new description for the 'Issue' discussion type in the UserTrendList component to improve clarity on user actions.
- Simplified the ContributorsRank component's score display logic for better maintainability.
- Removed unnecessary console log statements in the Toc component to streamline the code.

* fix: revoke logic

* fix: revoke logic

* fix: zero point not record

* fix: add logger

* fix: add logger

* fix: add logger

* fix: keyword disc search

* fix: accept disc can not delete by self

* fix: accept comment can not delete by comment user and disc user

* fix: remove data

* fix: user avatar delete

* fix: delete comment perm

* fix: delete disc perm

* fix: filter zero crontribute

* fix: filter zero crontribute

* fix: update bot comment

* fix: restore LLMTypeAnalysis and LLMTypeAnalysisVL enums; enhance ArticleSelector and CategorySelector with Chip rendering and textFieldSx prop; update ModelManagementModal for model configuration display; improve dashboard data aggregation for new metrics; refine UI components for better responsiveness and accessibility

---------

+++ backend/svc/discussion.go
@@ -46,6 +46,7 @@ type discussionIn struct {
 	Limiter        ratelimit.Limiter
 	LLM            *LLM
 	Batcher        batch.Batcher[model.StatInfo]
+	UserPoint      *UserPoint
 }
 
 type Discussion struct {
@@ -279,6 +280,10 @@ func (d *Discussion) Delete(ctx context.Context, user model.UserInfo, uuid strin
 
 	if !ok {
 		return errPermission
+	}
+
+	if disc.Type == model.DiscussionTypeQA && disc.Resolved == model.DiscussionStateResolved && user.Role != model.UserRoleAdmin && user.UID == disc.UserID {
+		return errors.New("resolved qa can not delete")
 	}
 
 	if err := d.in.DiscRepo.Delete(ctx, repo.QueryWithEqual("uuid", uuid)); err != nil {
@@ -564,6 +569,7 @@ type DiscussionKeywordAnswerReq struct {
 }
 
 func (d *Discussion) KeywordAnswer(ctx context.Context, req DiscussionKeywordAnswerReq) (string, error) {
+	logger := d.logger.WithContext(ctx).With("forum_id", req.ForumID).With("keyword", req.Keyword)
 	discs, err := d.Search(ctx, DiscussionSearchReq{
 		ForumID:             req.ForumID,
 		Keyword:             req.Keyword,
@@ -573,10 +579,12 @@ func (d *Discussion) KeywordAnswer(ctx context.Context, req DiscussionKeywordAns
 		return "", err
 	}
 
-	discIDs := make([]uint, len(discs))
+	discIDs := make([]uint, 0, len(discs))
 	for _, disc := range discs {
 		discIDs = append(discIDs, disc.ID)
 	}
+
+	logger.With("disc_ids", discIDs).Info("ai keyword answer get discs")
 
 	prompt, err := d.in.LLM.GenerateDiscussionPrompt(ctx, discIDs...)
 	if err != nil {
@@ -593,6 +601,7 @@ func (d *Discussion) KeywordAnswer(ctx context.Context, req DiscussionKeywordAns
 	}
 
 	if !answer {
+		logger.Info("ai can not answer with think")
 		return "", nil
 	}
 
@@ -707,9 +716,9 @@ func (d *Discussion) LikeDiscussion(ctx context.Context, discUUID string, user m
 			UserPointRecordInfo: model.UserPointRecordInfo{
 				UserID:    disc.UserID,
 				Type:      model.UserPointTypeLikeBlog,
-				ForeignID: user.UID,
+				ForeignID: disc.ID,
+				FromID:    user.UID,
 			},
-			FromUserID: user.UID,
 		})
 		if err != nil {
 			return err
@@ -761,7 +770,8 @@ func (d *Discussion) RevokeLikeDiscussion(ctx context.Context, discUUID string, 
 			UserPointRecordInfo: model.UserPointRecordInfo{
 				UserID:    disc.UserID,
 				Type:      model.UserPointTypeLikeBlog,
-				ForeignID: uid,
+				ForeignID: disc.ID,
+				FromID:    uid,
 			},
 			Revoke: true,
 		})
@@ -906,8 +916,8 @@ func (d *Discussion) CreateComment(ctx context.Context, uid uint, discUUID strin
 					UserID:    comment.UserID,
 					Type:      model.UserPointTypeAnswerQA,
 					ForeignID: comment.ID,
+					FromID:    disc.UserID,
 				},
-				FromUserID: disc.UserID,
 			})
 			if err != nil {
 				return 0, err
@@ -959,18 +969,20 @@ type CommentUpdateReq struct {
 }
 
 func (d *Discussion) UpdateComment(ctx context.Context, user model.UserInfo, discUUID string, commentID uint, req CommentUpdateReq) error {
-	if !d.allow("comment", discUUID, user.UID) {
-		return errRatelimit
-	}
-
 	disc, err := d.in.DiscRepo.GetByUUID(ctx, discUUID)
 	if err != nil {
 		return err
 	}
 
-	err = d.CheckPerm(ctx, user.UID, disc)
-	if err != nil {
-		return err
+	if !req.Bot {
+		if !d.allow("comment", discUUID, user.UID) {
+			return errRatelimit
+		}
+
+		err = d.CheckPerm(ctx, user.UID, disc)
+		if err != nil {
+			return err
+		}
 	}
 
 	comment, err := d.GetCommentByID(ctx, commentID)
@@ -1014,11 +1026,22 @@ func (d *Discussion) DeleteComment(ctx context.Context, user model.UserInfo, dis
 	if !user.CanOperator(comment.UserID) {
 		return errors.New("not allowed to delete comment")
 	}
+
+	if disc.Type == model.DiscussionTypeQA && comment.Accepted && user.Role != model.UserRoleAdmin && (comment.UserID == user.UID || disc.UserID == user.UID) {
+		return errors.New("accept comment can not delete")
+	}
+
 	if err := d.in.CommRepo.Delete(ctx, repo.QueryWithEqual("id", commentID)); err != nil {
 		return err
 	}
-	if err := d.in.CommLikeRepo.Delete(ctx, repo.QueryWithEqual("comment_id", commentID)); err != nil {
-		return err
+	if comment.Accepted {
+		err = d.in.DiscRepo.Update(ctx, map[string]any{
+			"resolved":    model.DiscussionStateNone,
+			"resolved_at": gorm.Expr("null"),
+		}, repo.QueryWithEqual("id", disc.ID))
+		if err != nil {
+			return err
+		}
 	}
 	d.in.Pub.Publish(ctx, topic.TopicCommentChange, topic.MsgCommentChange{
 		OP:       topic.OPDelete,
@@ -1027,6 +1050,15 @@ func (d *Discussion) DeleteComment(ctx context.Context, user model.UserInfo, dis
 		DiscID:   disc.ID,
 		DiscUUID: discUUID,
 	})
+	if disc.Type == model.DiscussionTypeQA && comment.ParentID == 0 {
+		err = d.in.UserPoint.RevokeCommentPoint(ctx, disc.ID, disc.UserID, comment)
+		if err != nil {
+			return err
+		}
+	}
+	if err := d.in.CommLikeRepo.Delete(ctx, repo.QueryWithEqual("comment_id", commentID)); err != nil {
+		return err
+	}
 	return nil
 }
 
@@ -1100,6 +1132,7 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 	if comment.Accepted {
 		err = d.in.CommRepo.Update(ctx, map[string]any{
 			"accepted":    false,
+			"accepted_by": 0,
 			"accepted_at": gorm.Expr("null"),
 		}, repo.QueryWithEqual("id", commentID))
 		if err != nil {
@@ -1111,6 +1144,7 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 				UserID:    comment.UserID,
 				Type:      model.UserPointTypeAnswerAccepted,
 				ForeignID: commentID,
+				FromID:    comment.AcceptedBy,
 			},
 			Revoke: true,
 		})
@@ -1131,6 +1165,7 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 				UserID:    disc.UserID,
 				Type:      model.UserPointTypeAcceptAnswer,
 				ForeignID: disc.ID,
+				FromID:    comment.AcceptedBy,
 			},
 			Revoke: true,
 		})
@@ -1151,6 +1186,7 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 
 	err = d.in.CommRepo.Update(ctx, map[string]any{
 		"accepted":    false,
+		"accepted_by": 0,
 		"accepted_at": gorm.Expr("null"),
 	}, repo.QueryWithEqual("discussion_id", disc.ID),
 		repo.QueryWithEqual("accepted", true))
@@ -1164,6 +1200,7 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 				UserID:    comm.UserID,
 				Type:      model.UserPointTypeAnswerAccepted,
 				ForeignID: comm.ID,
+				FromID:    comm.AcceptedBy,
 			},
 			Revoke: true,
 		})
@@ -1176,6 +1213,7 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 
 	if err := d.in.CommRepo.UpdateByModel(ctx, &model.Comment{
 		Accepted:   true,
+		AcceptedBy: user.UID,
 		AcceptedAt: model.Timestamp(now.Unix()),
 	}, repo.QueryWithEqual("id", commentID)); err != nil {
 		return err
@@ -1186,8 +1224,8 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 			UserID:    comment.UserID,
 			Type:      model.UserPointTypeAnswerAccepted,
 			ForeignID: commentID,
+			FromID:    user.UID,
 		},
-		FromUserID: user.UID,
 	})
 	if err != nil {
 		return err
@@ -1210,14 +1248,14 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 		}
 
 		// 自己的问题自己采纳回答
-		if user.UID == disc.UserID {
+		if user.UID == disc.UserID && disc.Type == model.DiscussionTypeQA {
 			err = d.in.Pub.Publish(ctx, topic.TopicUserPoint, topic.MsgUserPoint{
 				UserPointRecordInfo: model.UserPointRecordInfo{
 					UserID:    disc.UserID,
 					Type:      model.UserPointTypeAcceptAnswer,
 					ForeignID: disc.ID,
+					FromID:    user.UID,
 				},
-				FromUserID: comment.UserID,
 			})
 			if err != nil {
 				return err
@@ -1229,6 +1267,7 @@ func (d *Discussion) AcceptComment(ctx context.Context, user model.UserInfo, dis
 				UserID:    disc.UserID,
 				Type:      model.UserPointTypeAcceptAnswer,
 				ForeignID: disc.ID,
+				FromID:    disc.UserID,
 			},
 			Revoke: true,
 		})
@@ -1314,6 +1353,7 @@ func (d *Discussion) LikeComment(ctx context.Context, userInfo model.UserInfo, d
 					UserID:    comment.UserID,
 					Type:      model.UserPointTypeAnswerDisliked,
 					ForeignID: commentID,
+					FromID:    userInfo.UID,
 				},
 				Revoke: true,
 			})
@@ -1327,6 +1367,7 @@ func (d *Discussion) LikeComment(ctx context.Context, userInfo model.UserInfo, d
 						UserID:    userInfo.UID,
 						Type:      model.UserPointTypeDislikeAnswer,
 						ForeignID: commentID,
+						FromID:    comment.UserID,
 					},
 					Revoke: true,
 				})
@@ -1341,8 +1382,8 @@ func (d *Discussion) LikeComment(ctx context.Context, userInfo model.UserInfo, d
 				UserID:    comment.UserID,
 				Type:      model.UserPointTypeAnswerLiked,
 				ForeignID: commentID,
+				FromID:    userInfo.UID,
 			},
-			FromUserID: userInfo.UID,
 		})
 		if err != nil {
 			return err
@@ -1399,6 +1440,7 @@ func (d *Discussion) DislikeComment(ctx context.Context, userInfo model.UserInfo
 					UserID:    comment.UserID,
 					Type:      model.UserPointTypeAnswerLiked,
 					ForeignID: commentID,
+					FromID:    userInfo.UID,
 				},
 				Revoke: true,
 			})
@@ -1412,8 +1454,8 @@ func (d *Discussion) DislikeComment(ctx context.Context, userInfo model.UserInfo
 				UserID:    comment.UserID,
 				Type:      model.UserPointTypeAnswerDisliked,
 				ForeignID: commentID,
+				FromID:    userInfo.UID,
 			},
-			FromUserID: userInfo.UID,
 		})
 		if err != nil {
 			return err
@@ -1425,8 +1467,8 @@ func (d *Discussion) DislikeComment(ctx context.Context, userInfo model.UserInfo
 					UserID:    userInfo.UID,
 					Type:      model.UserPointTypeDislikeAnswer,
 					ForeignID: commentID,
+					FromID:    comment.UserID,
 				},
-				FromUserID: comment.UserID,
 			})
 			if err != nil {
 				return err
@@ -1475,6 +1517,7 @@ func (d *Discussion) RevokeLike(ctx context.Context, uid uint, discUUID string, 
 					UserID:    comment.UserID,
 					Type:      model.UserPointTypeAnswerDisliked,
 					ForeignID: commentID,
+					FromID:    commentLike.UserID,
 				},
 				Revoke: true,
 			})
@@ -1488,6 +1531,7 @@ func (d *Discussion) RevokeLike(ctx context.Context, uid uint, discUUID string, 
 						UserID:    commentLike.UserID,
 						Type:      model.UserPointTypeDislikeAnswer,
 						ForeignID: commentID,
+						FromID:    comment.UserID,
 					},
 					Revoke: true,
 				})
@@ -1502,6 +1546,7 @@ func (d *Discussion) RevokeLike(ctx context.Context, uid uint, discUUID string, 
 					UserID:    comment.UserID,
 					Type:      model.UserPointTypeAnswerLiked,
 					ForeignID: commentID,
+					FromID:    commentLike.UserID,
 				},
 				Revoke: true,
 			})
@@ -1714,12 +1759,22 @@ func (d *Discussion) AssociateDiscussion(ctx context.Context, user model.UserInf
 			UserID:    disc.UserID,
 			Type:      model.UserPointTypeAssociateIssue,
 			ForeignID: disc.ID,
+			FromID:    user.UID,
 		},
-		FromUserID: user.UID,
 	})
 	if err != nil {
 		return err
 	}
 
 	return nil
+}
+
+func (d *Discussion) GetBotComment(ctx context.Context, discID uint) (*model.Comment, error) {
+	var res model.Comment
+	err := d.in.CommRepo.GetBotComment(ctx, &res, discID)
+	if err != nil {
+		return nil, err
+	}
+
+	return &res, nil
 }
